+++ src/charts/specs.js
@@ -1,0 +1,23 @@
+import bar_horizontal from './bar-horizontal.json';
+import bar from './bar.json';
+import bubble from './bubble.json';
+import grouped from './grouped.json';
+import pie from './pie.json';
+import scatter from './scatter.json';
+import sparkline from './sparkline.json';
+import time_trendline from './time-trendline.json';
+import time from './time.json';
+
+const specTemplates = {
+  bar_horizontal,
+  bar,
+  bubble,
+  grouped,
+  pie,
+  scatter,
+  sparkline,
+  time_trendline,
+  time
+};
+
+export default specTemplates;